requirements.txt
django==4.1.0
django-environ==0.9.0
pillow==9.4.0 
django-taggit==3.1.0 
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-allauth==0.50.0
coverage==7.0.5 
 django-admin-honeypot==1.1.0
 whitenoise==6.1.0
 gunicorn==20.1.0

